refactor(autoware_tensorrt_plugins): split tests into logical units
